+++ src/Exceptions.class.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * File that holds all our 
+ * Exception classes.
+ *
+ * @class General_Exception @extends $Exception
+ * @class Pythia_Exception  @extends General_Exception
+ *
+ */
+class General_Exception extends Exception
+{
+}
+
+class Pythia_Exception extends General_Exception
+{
+}
+?>